requirements.txt
torch==2.10.0
torchvision
segmentation-models-pytorch
albumentations
wandb
opencv-python
numpy
pyyaml
torchmetrics
matplotlib
tqdm